Allowed everything to be translated.
This can be done via `site.php`, `app.php` or `spoon.php` - whichever is most appropriate for the text. For example Tab and Group names will be in `site.php`, core text strings will be in `app.php` and anything related to the Spoon interface will be in `spoon.php`.

Fixes #22.